roles for user and addeduser

+++ app/Http/Controllers/Api/AddedUserController.php
@@ -2,26 +2,28 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Http\Controllers\Controller;
-use App\Http\Requests\StoreAddedUserRequest;
-use App\Http\Resources\AddedUserResource;
-use App\Http\Resources\CountryResource;
-use App\Models\AddedUser;
-use App\Models\Attachment;
+use Carbon\Carbon;
 use App\Models\Country;
 use App\Services\Search;
-use App\Traits\AttachPhotosTrait;
-use Carbon\Carbon;
-use Illuminate\Http\Request;
-use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use App\Models\AddedUser;
+use App\Models\Attachment;
 use Illuminate\Support\Arr;
+use Illuminate\Http\Request;
+use App\Traits\AttachPhotosTrait;
 use Illuminate\Support\Facades\DB;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use App\Http\Resources\CountryResource;
+use App\Http\Resources\AddedUserResource;
+use App\Http\Requests\StoreAddedUserRequest;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+
 class AddedUserController extends Controller
 {
     use AttachPhotosTrait;
     public function __construct(private Search $search)
     {
-        $this->authorizeResource(AddedUser::class);
+        // $this->authorizeResource(AddedUser::class);
         $this->search = new Search();
     }
 
@@ -32,13 +34,31 @@ class AddedUserController extends Controller
      */
     public function index(Request $request, Country $country)
     {
+        $this->authorize('viewAny', AddedUser::class);
+        
+        $limit = 100;
+    
         if (isset($country['id'])) {
-            return AddedUserResource::collection($country->addedUsers);
+            $addedUsers = $country->addedUsers()->paginate($limit);
+        } elseif ($request->has('risk')) {
+            $addedUsers = AddedUser::with(['country'])
+                ->where('sanction', $request->get('risk'))
+                ->orderBy('created_at', 'desc')
+                ->paginate($limit);
+        } else {
+            $addedUsers = AddedUser::with(['country'])
+                ->orderBy('created_at', 'desc')
+                ->paginate($limit);
         }
-        if ($request->has('risk')) {
-            return AddedUserResource::collection(AddedUser::with(['country'])->where('sanction', $request->get('risk'))->orderBy('created_at', 'desc')->get());
-        }
-        return AddedUserResource::collection(AddedUser::with(['country'])->orderBy('created_at', 'desc')->get());
+        
+        $page = AddedUserResource::collection($addedUsers);
+        
+        return response()->json([
+            $page->items(),
+            ['previousPageUrl' => $page->previousPageUrl(),
+            'nextPageUrl' => $page->nextPageUrl(),
+            'totalPages' => $page->lastPage(),]
+        ]);
     }
 
     /**
@@ -49,27 +69,31 @@ class AddedUserController extends Controller
      */
     public function store(StoreAddedUserRequest $request)
     {
-            $user = DB::transaction(function () use ($request) {
-                $user = AddedUser::create(
-                    Arr::except($request->validated(), ['passport_photo', 'cv_photo'])
-                );
-                $this->attachPhotos($request, $user);
-                return $user;
-            });
+        $this->authorize('create',AddedUser::class);
 
-            return new AddedUserResource($user);
+        $user = DB::transaction(function () use ($request) {
+            $user = AddedUser::create(
+                Arr::except($request->validated(), ['passport_photo', 'cv_photo'])
+            );
+            $this->attachPhotos($request, $user);
+            return $user;
+        });
+
+        return new AddedUserResource($user);
     }
 
 
 
     public function delete(Attachment $attachment)
     {
+        $this->authorize('delete',AddedUser::class);
         $attachment->delete();
         return response()->json([], 204);
     }
 
     public function upload(Request $request, AddedUser $addedUser)
     {
+        $this->authorize('create',$addedUser);
         $request->validate([
             'passport_photo.*' => 'image',
             'cv_photo.*' => 'image',
@@ -97,6 +121,7 @@ class AddedUserController extends Controller
      */
     public function show(AddedUser $addedUser)
     {
+        $this->authorize('view',AddedUser::class);
         return new AddedUserResource($addedUser->loadMissing('userOperations'));
     }
 
@@ -109,6 +134,7 @@ class AddedUserController extends Controller
      */
     public function update(StoreAddedUserRequest $request, AddedUser $addedUser)
     {
+        $this->authorize('update',$addedUser);
         $addedUser->update(
             Arr::except($request->validated(), ['passport_photo', 'cv_photo'])
         );
@@ -127,6 +153,7 @@ class AddedUserController extends Controller
      */
     public function destroy(AddedUser $addedUser)
     {
+        $this->authorize('delete',$addedUser);
         $addedUser->delete();
         return response()->noContent();
     }
